requirements.txt
langchain==0.1.16
langchain-community==0.0.34
chromadb==0.4.24
sentence-transformers==2.7.0
transformers==4.41.2
torch==2.3.0
accelerate==0.29.3
python-dotenv==1.0.1
googlesearch-python==1.2.3
numpy>=1.21.0,<2.0.0
chromadb==0.4.22
sentence-transformers>=2.2.2
